organized svgs into folders
