+++ src/pages/get-involved.js
@@ -16,8 +16,8 @@ import {
   Image2,
   PositionCardsContainer,
 } from "../page-styles/get-involved-styles"
-import top_svg from "../images/getinvolved_top_svg.svg"
-import middle_svg from "../images/getinvolved_middle_svg.svg"
+import top_svg from "../images/get-involved/top_shape.svg"
+import middle_svg from "../images/get-involved/middle_shape.svg"
 
 const GetInvolved = props => {
   const involvedPage = get(props, "data.contentfulGetInvolvedPage")
